requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-login==1.10.3
fastapi-users==14.0.1
greenlet==3.2.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
makefun==1.15.6
MarkupSafe==3.0.2
packaging==24.2
paho-mqtt==2.1.0
pluggy==1.5.0
pwdlib==0.2.1
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
PyJWT==2.10.1
pytest==8.3.5
python-multipart==0.0.20
python-telegram-bot==22.0
PyYAML==6.0.2
setuptools==78.1.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.0
uvicorn==0.34.0
pillow==11.2.1
